rename AutoCloseableSupplier -> ReloadingCloseableContainer, and ReloadingCloseableContainer -> ReloadingCloseableContainerImpl
