--- atlasdb-cassandra/src/main/java/com/palantir/atlasdb/cassandra/AutoCloseableSupplier.java
@@ -1,23 +1,0 @@
-/*
- * (c) Copyright 2022 Palantir Technologies Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.palantir.atlasdb.cassandra;
-
-import java.util.function.Supplier;
-
-public interface AutoCloseableSupplier<T extends AutoCloseable> extends AutoCloseable, Supplier<T> {
-    boolean isClosed();
-}
